requirements.txt
# Core Framework
fastapi==0.104.1
streamlit==1.28.1
uvicorn==0.24.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# Analytics & Sentiment
textblob==0.17.1
vaderSentiment==3.3.2

# Visualization
plotly==5.17.0
plotly-express==0.4.1

# Task Scheduling
schedule==1.2.0

# Utilities  
python-dateutil==2.8.2
pytz==2023.3